backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
Pillow
python-multipart
